fix failing test
