+++ src/boundViews/__tests__/classes/complexView.ts
