+++ src/createElements/_typeGuards.ts
@@ -1,4 +1,4 @@
-import { IElemDefinition, IDrawable, DrawableElement, IClassDefinition, ClassName } from "./_interfaces";
+import { IElemDefinition, IClassDefinition, ClassName } from "./_interfaces";
 import { isInterface, isString, isArray } from "../shared";
 
 /** 
@@ -23,16 +23,6 @@ export function isIElemDefinition(test: any): test is IElemDefinition {
     if (isInterface<IElemDefinition>(test, comp)) { return true; }
     return false;
 
-}
-
-/** check if the element implements the drawable class */
-export function isDrawable(test: any): test is IDrawable {
-    return !!(test as IDrawable).draw;
-}
-
-/** check if the element is one that can be used as a drawable base */
-export function isDrawableElement(test: any): test is DrawableElement {
-    return (!!(test.appendChild));
 }
 
 export function isClassDefinition(test: IClassDefinition | ClassName): test is IClassDefinition {
