+++ src/shared/_interfaces.ts
@@ -1,6 +1,19 @@
 export type StandardElement = HTMLElement | SVGElement;
 
 /**
+ * IDrawable
+ * ----------------------------------------------------------------------------
+ * The core pieecs that are required by our Drawable
+ */
+export interface IDrawable {
+    draw(parent?: DrawableElement): void;
+    erase(): void;
+    base: StandardElement;
+}
+
+export type DrawableElement = StandardElement | IDrawable;
+
+/**
  * IPoint
  * ---------------------------------------------------------------------------
  * Defines a basic point in space
@@ -10,3 +23,4 @@ export interface IPoint {
 	y: number;
 	z?: number;
 }
+
